+++ platform/mellanox/docker-syncd-mlnx/ecmp_calculator/ecmp_calc_sdk.py
@@ -1,15 +1,10 @@
 #!/usr/bin/env python3
 
 from python_sdk_api.sx_api import *
+from port_utils import sx_check_rc
 import struct
 import socket
 import inspect
-
-DEVICE_ID = 1
-SWITCH_ID = 0
-ETHERNET_PREFIX = 'Ethernet'
-ASIC_MAX_LANES = {SX_CHIP_TYPE_SPECTRUM: 4, SX_CHIP_TYPE_SPECTRUM2: 4, 
-                  SX_CHIP_TYPE_SPECTRUM3: 8, SX_CHIP_TYPE_SPECTRUM4: 8}
 
 def sx_open_sdk_connection():
     """ Open connection to SDK
@@ -25,35 +20,6 @@ def sx_open_sdk_connection():
     
     return handle
     
-def sx_get_chip_type(handle):
-    """ Get system ASIC type
-    
-    Args:
-        handle (sx_api_handle_t): SDK handle
-    
-    Returns:
-        sx_chip_types_t : Chip type
-    """       
-    try:
-        device_info_cnt_p = new_uint32_t_p()
-        uint32_t_p_assign(device_info_cnt_p, 1)
-        device_info_cnt = uint32_t_p_value(device_info_cnt_p)
-        device_info_list_p = new_sx_device_info_t_arr(device_info_cnt)
-        
-        rc = sx_api_port_device_list_get(handle, device_info_list_p, device_info_cnt_p)
-        sx_check_rc(rc)
-    
-        device_info = sx_device_info_t_arr_getitem(device_info_list_p, 0)
-        chip_type = device_info.dev_type
-        if chip_type == SX_CHIP_TYPE_SPECTRUM_A1:
-            chip_type = SX_CHIP_TYPE_SPECTRUM
-            
-        return chip_type
-    
-    finally:
-        delete_uint32_t_p(device_info_cnt_p)
-        delete_sx_device_info_t_arr(device_info_list_p)
-
 def sx_get_active_vrids(handle):
     """ Get existing virtual router IDs
     
@@ -89,58 +55,9 @@ def sx_get_active_vrids(handle):
         return vrid_list
     
     finally:
-        delete_uint32_t_p(vrid_cnt_p)
-        delete_sx_router_id_t_p(vrid_key_p)
         delete_sx_router_id_t_arr(vrid_list_p)
-
-def sx_get_ports_map(handle):
-    """ Get ports map from SDK logical index to SONiC index
-    
-    Args:
-        handle (sx_api_handle_t): SDK handle
-    
-    Returns:
-        dict : Dictionary of ports indices. Key is SDK logical index, value is SONiC index (4 for Ethernet4)
-    """         
-    try:
-        ports_map = {}
-        
-        # Get chip type
-        chip_type = sx_get_chip_type(handle)
-        
-        # Get ports count
-        port_cnt_p = new_uint32_t_p()
-        rc = sx_api_port_device_get(handle, DEVICE_ID, SWITCH_ID, None,  port_cnt_p)
-        sx_check_rc(rc)
-        
-        # Get ports
-        port_cnt = uint32_t_p_value(port_cnt_p)
-        port_attributes_list = new_sx_port_attributes_t_arr(port_cnt)
-        rc = sx_api_port_device_get(handle, DEVICE_ID, SWITCH_ID, port_attributes_list,  port_cnt_p)
-        sx_check_rc(rc)
-        
-        for i in range(0, port_cnt):
-            port_attributes = sx_port_attributes_t_arr_getitem(port_attributes_list, i)
-            label_port = port_attributes.port_mapping.module_port
-            logical_port = port_attributes.log_port;
-            lane_bmap = port_attributes.port_mapping.lane_bmap;
-            
-            if (is_phys_port(port_attributes.log_port) == False):
-                continue
-            
-            # Calculate sonic index (sonic index=4 for Ethernet4)
-            lane_index = get_lane_index(lane_bmap, ASIC_MAX_LANES[chip_type])
-            assert lane_index != -1, "Failed to calculate port index"
-            
-            sonic_index = label_port * ASIC_MAX_LANES[chip_type] + lane_index;
-            sonic_interface = ETHERNET_PREFIX + str(sonic_index)    
-            ports_map[logical_port] = sonic_interface
-            
-        return ports_map
-    
-    finally:
-        delete_uint32_t_p(port_cnt_p)
-        delete_sx_port_attributes_t_arr(port_attributes_list)
+        delete_sx_router_id_t_p(vrid_key_p)
+        delete_uint32_t_p(vrid_cnt_p)
 
 def sx_router_get_ecmp_id(handle, vrid, ip_prefix):
     """ Get ECMP id for a given IP prefix
@@ -172,9 +89,9 @@ def sx_router_get_ecmp_id(handle, vrid, ip_prefix):
         return 0
     
     finally:
-        delete_sx_ip_prefix_t_p(ip_prefix_p)
-        delete_uint32_t_p(entries_cnt_p)
         delete_sx_uc_route_get_entry_t_arr(entries_array)
+        delete_uint32_t_p(entries_cnt_p)
+        delete_sx_ip_prefix_t_p(ip_prefix_p)
 
 def sx_router_operational_ecmp_get(handle, ecmp_id):
     """ Get next hops for a given ECMP id
@@ -210,9 +127,9 @@ def sx_router_operational_ecmp_get(handle, ecmp_id):
         return next_hops
     
     finally:
-        delete_uint32_t_p(next_hop_count_p)
         delete_sx_next_hop_t_arr(next_hop_list_p)
-
+        delete_uint32_t_p(next_hop_count_p)
+        
 def sx_get_router_interface(handle, vrid, rif):
     """ Get router interface information
     
@@ -247,9 +164,9 @@ def sx_get_router_interface(handle, vrid, rif):
         return rif_params
         
     finally:
-        delete_sx_router_id_t_p(vrid_p)
+        delete_sx_interface_attributes_t_p(ifc_attr_p)
         delete_sx_router_interface_param_t_p(ifc_p)
-        delete_sx_interface_attributes_t_p(ifc_attr_p)    
+        delete_sx_router_id_t_p(vrid_p)
 
 def sx_port_vport_base_get(handle, vport):
     """ Get SDK logical index for given vport
@@ -273,9 +190,9 @@ def sx_port_vport_base_get(handle, vport):
         return logical_port  
       
     finally:
+        delete_sx_port_log_id_t_p(logical_port_p) 
         delete_sx_vlan_id_t_p(vlan_id_p)
-        delete_sx_port_log_id_t_p(logical_port_p)   
-        
+
 def sx_router_neigh_get_mac(handle, rif, addr, type_ipv6=False):
     """ Get neighbour mac address
     
@@ -307,9 +224,8 @@ def sx_router_neigh_get_mac(handle, rif, addr, type_ipv6=False):
     
     finally:
         delete_sx_neigh_filter_t_p(filter_p)
-        delete_uint32_t_p(neigh_entry_cnt_p)
         delete_sx_neigh_get_entry_t_arr(neigh_entry_list_p)    
-        
+        delete_uint32_t_p(neigh_entry_cnt_p)
         
 def sx_fdb_uc_mac_addr_get(handle, vlan_id, mac_addr):
     """ Get UC mac entry from FDB
@@ -329,7 +245,6 @@ def sx_fdb_uc_mac_addr_get(handle, vlan_id, mac_addr):
         key.entry_type = SX_FDB_UC_AGEABLE
         key.action = SX_FDB_ACTION_FORWARD
         key_p = copy_sx_fdb_uc_mac_addr_params_t_p(key)
-        
         key_filter = sx_fdb_uc_key_filter_t()
         key_filter.filter_by_fid =  SX_FDB_KEY_FILTER_FIELD_VALID
         key_filter.filter_by_mac_addr = SX_FDB_KEY_FILTER_FIELD_VALID
@@ -337,7 +252,6 @@ def sx_fdb_uc_mac_addr_get(handle, vlan_id, mac_addr):
         key_filter.fid = vlan_id
         key_filter.mac_addr = ether_addr(mac_addr)
         key_filter_p = copy_sx_fdb_uc_key_filter_t_p(key_filter)
-    
         data_cnt_p = copy_uint32_t_p(SX_FDB_MAX_GET_ENTRIES)
         mac_list_p = new_sx_fdb_uc_mac_addr_params_t_arr(SX_FDB_MAX_GET_ENTRIES)
         
@@ -353,10 +267,10 @@ def sx_fdb_uc_mac_addr_get(handle, vlan_id, mac_addr):
         return mac_entry
 
     finally:
-        delete_sx_fdb_uc_key_filter_t_p(key_filter_p)
-        delete_sx_fdb_uc_mac_addr_params_t_p(key_p)
         delete_sx_fdb_uc_mac_addr_params_t_arr(mac_list_p)
-        delete_uint32_t_p(data_cnt_p)       
+        delete_uint32_t_p(data_cnt_p)    
+        delete_sx_fdb_uc_key_filter_t_p(key_filter_p)
+        delete_sx_fdb_uc_mac_addr_params_t_p(key_p)   
         
 def sx_lag_port_group_get(handle, lag_id):
     """ Get LAG members
@@ -370,9 +284,10 @@ def sx_lag_port_group_get(handle, lag_id):
     """      
     try:
         lag_members = []
-        port_count_p = copy_uint32_t_p(0)
+        port_count_p = new_uint32_t_p()
+        uint32_t_p_assign(port_count_p, 0)
         port_arr = None
-    
+
         rc = sx_api_lag_port_group_get(handle, 0, lag_id, port_arr, port_count_p)
         sx_check_rc(rc)
         
@@ -388,17 +303,8 @@ def sx_lag_port_group_get(handle, lag_id):
         return lag_members
     
     finally:
-        delete_uint32_t_p(port_count_p)
         delete_sx_port_log_id_t_arr(port_arr) 
-                  
-def sx_check_rc(rc):
-    if rc is not SX_STATUS_SUCCESS:
-        # Get the calling function name from the last frame
-        cf = inspect.currentframe().f_back
-        func_name = inspect.getframeinfo(cf).function
-        error_info = func_name + ' failed with rc = ' + str(rc)
-    
-        raise Exception(error_info)
+        delete_uint32_t_p(port_count_p)
 
 def sx_make_ip_prefix_v4(addr, mask):
     """ Create IPv4 prefix
@@ -438,29 +344,4 @@ def sx_make_ip_prefix_v6(addr, mask):
         
     return ip_prefix
 
-def get_lane_index(lane_bmap, max_lanes):
-    """ Get index of first lane in use (2 for 00001100)
-    
-    Args:
-        lane_bmap (int): bitmap indicating module lanes in use
-        max_lanes (int): Max lanes in module
-
-    Returns:
-        int : index of the first bit set to 1 in lane_bmap
-    """       
-    for lane_idx in range(0, max_lanes):
-        if (lane_bmap & 0x1 == 1):
-            return lane_idx
-        lane_bmap = lane_bmap >> 1
-    return -1  
-
-def get_port_type(log_port_id):
-    return (log_port_id & SX_PORT_TYPE_ID_MASK) >> SX_PORT_TYPE_ID_OFFS
-
-def is_phys_port(log_port_id):
-    return get_port_type(log_port_id) == SX_PORT_TYPE_NETWORK
-
-def is_lag(log_port_id):
-    return get_port_type(log_port_id) == SX_PORT_TYPE_LAG
-
